requirements.txt
aiogram==3.27.0
aiosqlite==0.22.1
python-dotenv==1.2.2
apscheduler>=3.10,<4
